pyproject.toml
[project]
name = "sermon-splitter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "datasets>=4.4.1",
    "librosa>=0.11.0",
    "mediapipe>=0.10.14",
    "python-dotenv>=1.1.1",
    "opencv-python>=4.12.0.88",
    "soundfile>=0.13.1",
    "srt>=3.5.3",
    "streamlit>=1.50.0",
    "transformers>=4.57.1",
    "torch",
    "pre-commit>=4.3.0",
    "pytubefix>=10.3.4",
    "imageio-ffmpeg>=0.6.0",
    "ffmpeg-python>=0.2.0",
    "indic-transliteration>=2.3.75",
    "pytest>=9.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]

[tool.ruff]
exclude = ["deprecated"]
